requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
requests==2.31.0
python-decouple==3.8
groq==0.4.1
